+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Return appropriate error message when item for manual fee/fine can't be found. Refs UIU-2701.
 * *BREAKING* Add `pronouns` field to user edit form. Refs UIU-3119.
 * Add `pronouns` field to user view form. Refs UIU-3118.
+* Pronouns Field - add Character Limit Warning. Refs UIU-3324.
 * Add ability to create/edit role assignments for all of a user's affiliations. Refs UIU-3179.
 
 ## [11.0.11](https://github.com/folio-org/ui-users/tree/v11.0.11) (2025-01-15)